+++ src/services/index.ts
@@ -1,8 +1,12 @@
+export { CloudRunService } from './cloud-run.js';
 export * from './firebase-app.errors.js';
 export { FirebaseAppService, FirebaseAppType } from './firebase-app.js';
 export { FirebaseEmulatorService } from './firebase-emulator.js';
 export { GcloudEmulatorService } from './gcloud-emulator.js';
 export { GoogleApisService } from './google-apis.js';
+export { IamService } from './iam.js';
+export { PubSubService } from './pubsub.js';
+export { ResourceManagerService } from './resource-manager.js';
 export { GoogleSecretManagerService } from './secret-manager.js';
 export * from './storage.errors.js';
 export { CloudStorageService } from './storage.js';